requirements.txt
groq>=1.1.2
sentence-transformers>=3.4.1
numpy>=1.26.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
rich>=13.7.0
httpx>=0.27.0